\name{betatgarch3-package}
\alias{betatgarch3-package}
\alias{betatgarch3}
\docType{package}
\title{\packageTitle{betatgarch3}}
\description{\packageDescription{betatgarch3}}
\section{Package Content}{\packageIndices{betatgarch3}}
\author{\packageAuthor{betatgarch3}}
\section{Maintainer}{\packageMaintainer{betatgarch3}}
\keyword{package}
