% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_class.R
\name{BetaTGARCHfit}
\alias{BetaTGARCHfit}
\title{BetaTGARCHfit: Estimation And Related Methods}
\description{
The class provides a method to estimate the Beta-t-GARCH(1,1)
model and access estimates, standard errors, etc.
}
\references{
\insertRef{Harvey2008}{betatgarch}
\insertRef{Blasques2018}{betatgarch}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{original data passed to BetaTGARCHfit}

\item{\code{matcoef}}{matrix of estimates, robust standard errors, robust t
statistics and associated p-values}

\item{\code{se}}{estimated standard errors}

\item{\code{nlopt_res}}{the raw results returned by \code{NLOPT}}

\item{\code{boolean}}{that indicates whether estimation was successful}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{boolean}}{that indicates whether estimation was successful}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BetaTGARCHfit$new()}}
\item \href{#method-nobs}{\code{BetaTGARCHfit$nobs()}}
\item \href{#method-fit}{\code{BetaTGARCHfit$fit()}}
\item \href{#method-filter}{\code{BetaTGARCHfit$filter()}}
\item \href{#method-logLik}{\code{BetaTGARCHfit$logLik()}}
\item \href{#method-coef}{\code{BetaTGARCHfit$coef()}}
\item \href{#method-vcov}{\code{BetaTGARCHfit$vcov()}}
\item \href{#method-residuals}{\code{BetaTGARCHfit$residuals()}}
\item \href{#method-sigma}{\code{BetaTGARCHfit$sigma()}}
\item \href{#method-show}{\code{BetaTGARCHfit$show()}}
\item \href{#method-summary}{\code{BetaTGARCHfit$summary()}}
\item \href{#method-clone}{\code{BetaTGARCHfit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{BetaTGARCHfit} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$new(y, n = 5L, f_0 = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{a numeric vector or univariate zoo/xts time series}

\item{\code{n}}{positive integer giving the size of the pre-sample}

\item{\code{f_0}}{optional, value of conditional variance with which to
initialise the filter. Default is \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The first \code{n} observations of \code{y} are used to estimate the
initial value of the conditional variance, \code{f_0}. The first
\code{n} observations also do not enter the likelihood calculation. If
\code{f_0} is given then the value of \code{n} is ignored.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nobs"></a>}}
\if{latex}{\out{\hypertarget{method-nobs}{}}}
\subsection{Method \code{nobs()}}{
Returns the number of observations used in estimation. This excludes the
\code{n} pre-sample observations used to initialise the filter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$nobs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Maximum Likelihood Estimation of The Beta-t-GARCH(1,1) Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$fit(
  x0 = NULL,
  restrict = TRUE,
  lb = c(.Machine$double.eps, 0, 0, -1, 2.1),
  ub = c(Inf, 1, 1, 1, Inf),
  opts = list(algorithm = "NLOPT_LD_AUGLAG", xtol_rel = 1e-12, maxeval = 1000,
    local_opts = list(algorithm = "NLOPT_LD_LBFGS", xtol_rel = 1e-10, maxeval = 500)),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x0}}{initial values passed to \code{nloptr}}

\item{\code{restrict}}{logical that indicates whether to optimise the model over
a restricted parameter space that satisfies an empirical version of the
Lyapunov condition.}

\item{\code{lb, ub}}{vectors that specify the upper and lower bounds for model
parameters}

\item{\code{opts}}{a list that specifies the algorithm and other options used by}

\item{\code{...}}{additional arguments passed to \code{nloptr}.
\code{nloptr} to minimise the negative log-likelihood.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function estimates the Beta-t-GARCH(1,1) model with leverage of
  \insertCite{Harvey2008;textual}{betatgarch}. It can, optionally, maximise
  the log-likelihood over a restricted parameter space that satisfies an
  empirical version of the Lyapunov condition as in
  \insertCite{Blasques2018;textual}{betatgarch}.

  NOTE: If you receive the error
  "\code{student_t_lpdf: Scale parameter is 0, but must be > 0!}"
  then try to increase the lower bound on the parameter nu.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter"></a>}}
\if{latex}{\out{\hypertarget{method-filter}{}}}
\subsection{Method \code{filter()}}{
Filter with the estimated model. This method will return a list with the
following elements:
  \enumerate{
    \item each observation's contribution to the log-likelihood (`"llik_t"`),
    \item the conditional variances (f),
    \item score residuals (s_t) and
    \item standardised residuals (residuals).
  }
Note that this method will update several other fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$filter(parm = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parm}}{parameter vector. If \code{NULL} (default) then it uses
estimated values. Supplying a parameter vector is not recommended since
\code{filter} will overwrite several fields of the object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logLik"></a>}}
\if{latex}{\out{\hypertarget{method-logLik}{}}}
\subsection{Method \code{logLik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$logLik()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coef"></a>}}
\if{latex}{\out{\hypertarget{method-coef}{}}}
\subsection{Method \code{coef()}}{
Returns a vector of estimated coefficients/parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$coef()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vcov"></a>}}
\if{latex}{\out{\hypertarget{method-vcov}{}}}
\subsection{Method \code{vcov()}}{
Calculates the robust variance-covariance matrix of a fitted
 Beta-t-GARCH(1,1) model. This method returns matrix of estimated
 covariances between model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$vcov(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-residuals"></a>}}
\if{latex}{\out{\hypertarget{method-residuals}{}}}
\subsection{Method \code{residuals()}}{
Returns a vector of standardised residuals from the
 estimated model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$residuals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sigma"></a>}}
\if{latex}{\out{\hypertarget{method-sigma}{}}}
\subsection{Method \code{sigma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$sigma()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summary of Estimation Results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHfit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
