% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_class.R
\name{BetaTGARCHsim}
\alias{BetaTGARCHsim}
\title{BetaTGARCHsim: Simulate from a Beta-t-GARCH(1, 1) model}
\value{
Both of the methods \code{generate} and \code{simulate} returns a list
  of vectors. The elements of the list are (in order):
  \enumerate{
    \item the simulated observations (\code{y_t}),
    \item the conditional variance (\code{f_t}),
    \item score innovations (\code{s_t}),
    \item the contributions to the log-likelihood (\code{llik_t}) and
    \item the series of innovations (\code{innovations}).
  }
}
\description{
The class provides methods to simulate from either an estimated
Beta-t-GARCH(1,1) model or a model specified via its coefficients and an
initial value for the conditional variance.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-f_0}{\code{BetaTGARCHsim$f_0()}}
\item \href{#method-coef}{\code{BetaTGARCHsim$coef()}}
\item \href{#method-new}{\code{BetaTGARCHsim$new()}}
\item \href{#method-generate}{\code{BetaTGARCHsim$generate()}}
\item \href{#method-simulate}{\code{BetaTGARCHsim$simulate()}}
\item \href{#method-clone}{\code{BetaTGARCHsim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-f_0"></a>}}
\if{latex}{\out{\hypertarget{method-f_0}{}}}
\subsection{Method \code{f_0()}}{
Initial value of the conditional variance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHsim$f_0()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coef"></a>}}
\if{latex}{\out{\hypertarget{method-coef}{}}}
\subsection{Method \code{coef()}}{
Returns a vector of estimated coefficients/parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHsim$coef()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a simulation object for a Beta-t-GARCH(1, 1) model where the
coefficients and the initial of the conditional variance are either
extracted from an estimated model object or supplied by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHsim$new(model = NULL, f_0 = NULL, coef = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{NULL} or a \code{BetaTGARCHfit} object}

\item{\code{f_0}}{\code{NULL} or the initial value of the conditional variance}

\item{\code{coef}}{\code{NULL} or the model coefficients}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If \code{model = NULL} then the user must supply values for \code{f_0}
and \code{coef}. Moreover, it is assumed that a supplied model was
estimated.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\if{latex}{\out{\hypertarget{method-generate}{}}}
\subsection{Method \code{generate()}}{
Generates observations from a given vector of innovations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHsim$generate(innovations, B = ceiling(0.1 * length(innovations)))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{innovations}}{vector of innovations}

\item{\code{B}}{an integer that gives the number of periods to discard at the
start of the generated series, i.e., it is the burn-in period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-simulate}{}}}
\subsection{Method \code{simulate()}}{
Simulates a specified number of observations from the specified model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHsim$simulate(n, B = ceiling(0.1 * n))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer that specifies the number of observation to simulate}

\item{\code{B}}{an integer that gives the number of periods to discard at the
start of the generated series, i.e., it is the burn-in period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaTGARCHsim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
