% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cnstr}
\alias{cnstr}
\title{Non-Linear Inequality Constraints}
\usage{
cnstr(y, theta)
}
\arguments{
\item{y}{numeric vector of observations \eqn{y_0, y_1, ...}}

\item{theta}{numeric vector of parameter values}
}
\value{
Returns a list that contains the value of the left-hand side
(`"objective"`) of each constraint and the Jacobian ("jacobian") of the
constraint.
}
\description{
Non-Linear Inequality Constraints
}
\details{
This function evaluates the left-hand side of the constraints
\enumerate{
  \item \eqn{-(\alpha + \gamma) \le 0} and
  \item \eqn{\frac{1}{n} \sum_{t=1}^{n} \log \Lambda_{t} \left( \theta \right) + \delta.}
 }
}
