% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elc}
\alias{elc}
\title{Contributions to The Empirical Lyapunov Condition}
\usage{
elc(y, theta)
}
\arguments{
\item{y}{numeric vector of observations \eqn{y_0, y_1, ...}}

\item{theta}{numeric vector of parameter values}
}
\value{
Returns a vector with the values \eqn{\log{\Lambda_t(\theta)}} for
  \eqn{t=1,2,\ldots}
}
\description{
Contributions to The Empirical Lyapunov Condition
}
\details{
Calculates \eqn{\log{\Lambda_t(\theta)}} for \eqn{t=1,2,\ldots}
}
