% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{fit}
\alias{fit}
\title{Fit a Beta-t-GARCH Model Using Maximum Likelihood Estimation}
\usage{
fit(
  x,
  n = 5,
  theta_0 = NULL,
  lb = c(.Machine$double.eps, 0, 0, 0, 2.1),
  ub = c(Inf, 1, 1, 1, Inf),
  constrain = TRUE,
  opts = list(algorithm = "NLOPT_LD_SLSQP", xtol_rel = 1e-04, maxeval = 100),
  ...
)
}
\arguments{
\item{x}{numeric vector of observations}

\item{n}{an optional integer that give the number of observation to use as
the pre-sample.}

\item{theta_0}{either NULL or a numerical vector that specifies the initial
guess used by `nloptr` to find optimal parameter estimates.}

\item{lb, ub}{numeric vectors that specify the upper and lower bounds used by
the optimisation routine.}

\item{constrain}{logical that indicates whether to optimise the model over a
restricted parameter space that satisfies an empirical version of the
Lyapunov condition.}

\item{opts}{a list that specifies the algorithm and other options used by
`nloptr` to minimise the negative log-likelihood.}

\item{...}{additional arguments to pass to `nloptr`.}
}
\value{
Returns a list that contains both the output from `nloptr` and the result
  of filtering the series using the estimates obtained from `nloptr`.
}
\description{
Fit a Beta-t-GARCH Model Using Maximum Likelihood Estimation
}
\details{
This function estimates the Beta-t-GARCH(1,1) model with leverage of
  \insertCite{Harvey2008;textual}{betatgarch}. It can, optionally, maximise
  the log-likelihood over a restricted parameter space that satisfies an
  empirical version of the Lyapunov condition as in
  \insertCite{Blasques2018;textual}{betatgarch}.

  NOTE: If you receive the error
  "\code{student_t_lpdf: Scale parameter is 0, but must be > 0!}"
  then try to increase the lower bound on the parameter nu.
}
\references{
\insertRef{Harvey2008}{betatgarch}

\insertRef{Blasques2018}{betatgarch}
}
