% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate}
\alias{generate}
\title{Generate/Simulate From a Beta-t-GARCH Model}
\usage{
generate(n, theta, f_0 = theta[1]/(1 - theta[3]), burnin = ceiling(n * 0.1))
}
\arguments{
\item{n}{number of observations to simulate}

\item{theta}{numeric vector of parameter values}

\item{f_0}{the initial value of the conditional variance}

\item{burnin}{the length of the burn-in period for the the simulation}
}
\value{
Returns a list of vectors. The elements of the list are (in order):
  \enumerate{
    \item the simulated observations (`"y_t"`),
    \item the conditional variance (`"f_t"`),
    \item score innovations (`"s_t"`),
    \item the contributions to the log-likelihood (`"llik_t"`) and
    \item the series of innovations (`"innovations"`).
  }
}
\description{
Generate/Simulate From a Beta-t-GARCH Model
}
\details{
Simulates observations from a Beta-t-GARCH model.
}
