% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nll}
\alias{nll}
\title{Negative log-likelihood and its gradient.}
\usage{
nll(y, f_0, theta)
}
\arguments{
\item{y}{numeric vector of observations \eqn{y_0, y_1, ...}}

\item{f_0}{initial value of the conditional variance}

\item{theta}{numeric vector of parameter values}
}
\value{
Returns a list that contains the negative log-likelihood (`"objective"`)
  and the gradient (`"gradient"`) of the negative log-likelihood w.r.t. the
  model parameters.
}
\description{
Negative log-likelihood and its gradient.
}
\details{
Calculates the negative log-likelihood of the Beta-GARCH(1, 1) model
  with leverage and its gradient w.r.t. to vector of parameters.
}
