% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recursion_lst}
\alias{recursion_lst}
\title{Beta-t-GARCH Recursion/Filter}
\usage{
recursion_lst(y, f_0, theta)
}
\arguments{
\item{y}{numeric vector of observations \eqn{y_0, y_1, ...}}

\item{f_0}{initial value of the conditional variance}

\item{theta}{numeric vector of parameter values}
}
\value{
Returns a list of vectors. The elements of the list are (in order):
  \enumerate{
    \item each observation's contribution to the log-likelihood (`"llik_t"`),
    \item the conditional variances (`"f"`),
    \item score residuals (`"s_t"`) and
    \item model residuals (`"residuals"`).
  }
}
\description{
Beta-t-GARCH Recursion/Filter
}
\details{
Calculates the values of
  \itemize{
    \item each observation's contribution to the log-likelihood,
    \item the conditional variance,
    \item score innovation/residual and
    \item model residuals.
  }
}
