% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_lst}
\alias{simulate_lst}
\title{Simulate Values From Beta-t-GARCH Model Given a Vector of Innovations}
\usage{
simulate_lst(e, f_0, theta)
}
\arguments{
\item{e}{numeric vector of zero-mean and unit-variance innovations}

\item{f_0}{initial value of the conditional variance}

\item{theta}{numeric vector of parameter values}
}
\value{
Returns a list of vectors. The elements of the list are (in order):
  \itemize{
   \item the simulated observations (`"y_t"`),
   \item the conditional variance (`"f_t"`),
   \item score innovations (`"s_t"`),
   \item the contributions to the log-likelihood (`"llik_t"`) and
   \item the original series of innovations (`"innovations"`).
  }
}
\description{
Simulate Values From Beta-t-GARCH Model Given a Vector of Innovations
}
\details{
Simulates observations from a Beta-t-GARCH model.
}
